# api/openapi.yaml
openapi: 3.1.0
info:
  title: Codex CPA Pool v2 API
  version: 2.0.0-dev
  description: >-
    Contract for the Go Admin, self-service and public configuration APIs.
    Authored response text uses Accept-Language (en or zh-CN), defaults to en, and
    declares Content-Language and Vary: Accept-Language. Codes, IDs and stored values
    are language independent. Clients must not parse display messages for control flow.
servers:
  - url: /
...
paths:
...
  /site-config.json:
    get:
      operationId: getPublicSiteConfiguration
      tags: [public-site]
      responses:
        "200":
          description: Browser-safe branding, accepted email suffixes and client export defaults
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicSiteConfiguration"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /admin/api/session:
    post:
      operationId: createAdminSession
      tags: [admin-session]
      parameters:
        - $ref: "#/components/parameters/ManagementKey"
      responses:
        "201":
          $ref: "#/components/responses/AdminSessionResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    get:
      operationId: getAdminSession
      tags: [admin-session]
      security:
        - adminSession: []
      responses:
        "200":
          $ref: "#/components/responses/AdminSessionResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      operationId: deleteAdminSession
      tags: [admin-session]
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Session revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoggedOutResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /admin/api/overview/summary:
    get:
      operationId: getAdminOverviewSummary
      tags: [admin-overview]
      security:
        - adminSession: []
      responses:
        "200":
          description: Secret-free control-plane totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverviewPayload"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /admin/api/overview/catalog:
    get:
      operationId: getAdminOverviewCatalog
      tags: [admin-overview]
      security:
        - adminSession: []
      responses:
        "200":
          description: Selector identities with current operational status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverviewCatalog"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /admin/api/overview/status:
    get:
      operationId: getAdminOverviewStatus
      tags: [admin-overview]
      security:
        - adminSession: []
      responses:
        "200":
          description: Runtime-bound overview counters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverviewStatusPayload"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /admin/api/onboarding:
    get:
      operationId: getAdminOnboarding
      tags: [admin-onboarding]
      security:
        - adminSession: []
      responses:
        "200":
          description: Secret-free first-run configuration status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingStatus"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /admin/api/accounts:
    get:
      operationId: listAdminAccounts
      tags: [admin-accounts]
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/UsageWindowQuery"
        - $ref: "#/components/parameters/UsageStartAtQuery"
        - $ref: "#/components/parameters/UsageEndAtQuery"
        - name: fresh
          in: query
          description: Request a single-flight official quota refresh when the persistent snapshot is old enough.
          schema: { type: string, enum: ["0", "1"], default: "0" }
      responses:
        "200":
          description: Account catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountCatalog"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: createAdminAccount
      tags: [admin-accounts]
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountCreateRequest"
      responses:
        "201":
          description: Account created, rendered, probed, and activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountCreateResponse"
        "503":
          $ref: "#/components/responses/AccountLifecycleUnavailableResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /admin/api/accounts/usage-breakdown:
    get:
      operationId: getAdminAccountUsageBreakdown
      tags: [admin-accounts]
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/AccountQuery"
        - $ref: "#/components/parameters/UsageWindowQuery"
        - $ref: "#/components/parameters/UsageStartAtQuery"
        - $ref: "#/components/parameters/UsageEndAtQuery"
      responses:
        "200":
          description: Account usage breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageBreakdown"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /admin/api/runtime/services:
    get:
      operationId: listAdminRuntimeServices
      tags: [admin-runtime]
      security:
        - adminSession: []
      responses:
        "200":
          description: Exact Compose-project service catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeServiceCatalog"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /admin/api/runtime/jobs:
    get:
      operationId: listAdminRuntimeJobs
      tags: [admin-runtime]
      security:
        - adminSession: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 60
            default: 30
      responses:
        "200":
          description: Recent bounded in-process runtime jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeJobCatalog"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: createAdminRuntimeJob
      tags: [admin-runtime]
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuntimeJobRequest"
      responses:
        "202":
          description: Runtime job accepted into the bounded pool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeJobSubmissionResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /admin/api/users:
    get:
      operationId: listAdminUsers
      tags: [admin-users]
      security:
        - adminSession: []
      parameters:
        - name: view
          in: query
          schema:
            type: string
            enum: [summary, members]
            default: summary
        - $ref: "#/components/parameters/UsageWindowQuery"
        - $ref: "#/components/parameters/UsageStartAtQuery"
        - $ref: "#/components/parameters/UsageEndAtQuery"
        - name: q
          in: query
          schema:
            type: string
        - name: team_id
          in: query
          schema:
            type: string
        - name: usage_state
          in: query
          schema:
            type: string
            enum: [all, used, unused]
            default: all
        - name: sort
          in: query
          schema:
            type: string
            enum: [email, requests, tokens, quota, last_used]
            default: tokens
        - name: direction
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: fresh
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            enum: [25, 50, 100]
            default: 50
      responses:
        "200":
          description: Paginated secret-free user catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCatalog"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: createAdminUser
      tags: [admin-users]
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: One-time user credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCreateResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /admin/api/users/detail:
    get:
      operationId: getAdminUserDetail
      tags: [admin-users]
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/EmailQuery"
        - $ref: "#/components/parameters/UsageWindowQuery"
        - $ref: "#/components/parameters/UsageStartAtQuery"
        - $ref: "#/components/parameters/UsageEndAtQuery"
      responses:
        "200":
          description: Lazy secret-free user account detail for the selected usage window
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetailResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /admin/api/users/usage-breakdown:
    get:
      operationId: getAdminUserUsageBreakdown
      tags: [admin-users]
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/EmailQuery"
        - $ref: "#/components/parameters/UsageWindowQuery"
        - $ref: "#/components/parameters/UsageStartAtQuery"
        - $ref: "#/components/parameters/UsageEndAtQuery"
      responses:
        "200":
          description: User usage breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageBreakdown"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /admin/api/teams:
    get:
      operationId: listAdminTeams
      tags: [admin-teams]
      security:
        - adminSession: []
      responses:
        "200":
          description: Team catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamCatalog"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: createAdminTeam
      tags: [admin-teams]
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Created team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMutationResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    put:
      operationId: updateAdminTeam
      tags: [admin-teams]
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdateInput"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMutationResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      operationId: deleteAdminTeam
      tags: [admin-teams]
      security:
        - adminSession: []
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Deleted team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMutationResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /admin/api/settings/general:
    get:
      operationId: getAdminGeneralSettings
      tags: [admin-settings]
      security:
        - adminSession: []
      responses:
        "200":
          description: General settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralSettings"
        default:
          $ref: "#/components/responses/ErrorResponse"
    put:
      operationId: updateAdminGeneralSettings
      tags: [admin-settings]
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneralSettingsUpdateRequest"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralSettingsMutationResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /admin/api/settings/workspace:
    get:
      operationId: getAdminSettingsWorkspace
      tags: [admin-settings]
      security:
        - adminSession: []
      responses:
        "200":
          description: Configuration workspace metadata without secret contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsWorkspace"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /admin/api/settings/notifications:
    get:
      operationId: getAdminNotificationSettings
      tags: [admin-settings]
      security:
        - adminSession: []
      responses:
        "200":
          description: Notification settings and status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSettings"
        default:
          $ref: "#/components/responses/ErrorResponse"
    put:
      operationId: updateAdminNotificationSettings
      tags: [admin-settings]
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationSettingsUpdateRequest"
      responses:
        "200":
          description: Updated notification settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSettingsMutationResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /admin/api/overview/usage:
    get:
      operationId: getAdminOverviewUsage
      tags: [admin-overview]
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/UsageWindowQuery"
        - $ref: "#/components/parameters/UsageStartAtQuery"
        - $ref: "#/components/parameters/UsageEndAtQuery"
        - name: account
          in: query
          schema: { type: array, items: { type: string } }
        - name: user
          in: query
          schema: { type: array, items: { type: string, format: email } }
        - name: user_limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 10 }
        - name: token_mode
          in: query
          schema: { type: string, enum: [unweighted, weighted], default: unweighted }
      responses:
        "200":
          description: Fine-grained Token time series loaded only for the requested window and filters
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OverviewUsageResponse" }
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /admin/api/images/cliproxy:
    get:
      operationId: getAdminCPAImageStatus
      tags: [admin-runtime]
      security:
        - adminSession: []
      responses:
        "200":
          description: Applied, candidate, local and per-account CLIProxyAPI image identities
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CPAImageStatus" }
        default:
          $ref: "#/components/responses/ErrorResponse"
  /admin/api/release:
    get:
      operationId: getAdminReleaseStatus
      tags: [admin-runtime]
      security:
        - adminSession: []
      parameters:
        - name: fresh
          in: query
          schema: { type: string, enum: ["0", "1"], default: "0" }
      responses:
        "200":
          description: Current deployment and latest GitHub Release semantic versions
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReleaseStatus" }
        default:
          $ref: "#/components/responses/ErrorResponse"
  /admin/api/teams/usage:
    get:
      operationId: getAdminTeamUsage
      tags: [admin-teams]
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/UsageWindowQuery"
        - $ref: "#/components/parameters/UsageStartAtQuery"
        - $ref: "#/components/parameters/UsageEndAtQuery"
      responses:
        "200":
          description: Current-membership team totals for the requested range
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeamUsageResponse" }
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /admin/api/users/quota-actions:
    get:
      operationId: getAdminUserQuotaOperations
      tags: [admin-users]
      security:
        - adminSession: []
      responses:
        "200":
          description: Current-week all-user quota impact used before destructive bulk actions
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserQuotaOperationSummary" }
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: applyAdminUserQuotaAction
      tags: [admin-users]
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserQuotaActionRequest" }
      responses:
        "200":
          description: Bulk current-week quota policy or adjustment result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserQuotaActionResponse" }
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /admin/api/settings/configuration:
    get:
      operationId: getAdminConfiguration
      tags: [admin-settings]
      security:
        - adminSession: []
      responses:
        "200":
          description: Complete effective configuration catalog, including model and reasoning-effort quota multipliers, with presentation metadata and write-only secrets masked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigurationCatalog" }
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: updateAdminConfiguration
      tags: [admin-settings]
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConfigurationUpdateRequest" }
      responses:
        "200":
          description: Atomic settings and encrypted-proxy update with compensated runtime apply; multiplier changes affect only newly collected usage events
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigurationUpdateResponse" }
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /admin/api/notifications/test:
    post:
      operationId: testAdminNotification
      tags: [admin-settings]
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: One bounded channel-test message without collecting account quota or usage data
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SendNotificationResponse" }
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /admin/api/jobs:
    get:
      operationId: listLegacyAdminRuntimeJobs
      tags: [admin-runtime]
      security:
        - adminSession: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 60, default: 30 }
      responses:
        "200":
          description: Compatibility alias for recent runtime jobs
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LegacyRuntimeJobCatalog" }
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: cpa_admin_session
    portalSession:
      type: apiKey
      in: cookie
      name: cpa_usage_session
  parameters:
    ManagementKey:
      name: X-Management-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 12
        maxLength: 128
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 16
    AccountQuery:
      name: account
      in: query
      required: true
      schema:
        type: string
...
  responses:
...
    ErrorResponse:
      description: Safe JSON error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    AdminSessionResponse:
      description: Authenticated Admin session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AdminSession"
...
  schemas:
...
    PublicSiteConfiguration:
      type: object
      additionalProperties: false
      required:
        - version
        - timezone
        - product_name
        - short_name
        - environment_label
        - public_base_url
        - allowed_email_domains
        - provider_name
        - api_key_env
        - default_model
        - logo
      properties:
        version: { type: integer, const: 1 }
        timezone:
          type: string
          description: IANA timezone configured by system.timezone for all business dates and displays
        product_name: { type: string }
        short_name: { type: string }
        environment_label: { type: string }
        public_base_url: { type: string }
        allowed_email_domains:
          type: array
          description: Public email suffixes accepted by user creation and Usage Center login
          items: { type: string }
        provider_name: { type: string }
        api_key_env: { type: string }
        default_model: { type: string }
        logo: { $ref: "#/components/schemas/PublicSiteLogo" }
...
    LoggedOutResponse:
      type: object
      additionalProperties: false
      required: [logged_out]
      properties:
        logged_out:
          type: boolean
          const: true
...
    OverviewPayload:
      type: object
      additionalProperties: false
      required: [generated_at, source, summary]
      properties:
        generated_at:
          type: integer
          format: int64
        source:
          type: string
          const: control-plane
        summary:
          $ref: "#/components/schemas/OverviewSummary"
...